Modify PersonalInformation entity to include SemesterGrade relationship
I have modified the PersonalInformation entity to include a relationship field for SemesterGrade, in order to store the relationship between PersonalInformation and SemesterGrade.

Entity: I have added a one-to-many relationship field to the PersonalInformation entity to reference multiple SemesterGrade objects.
The modification to the PersonalInformation entity will allow for storing the relationship between PersonalInformation and multiple SemesterGrade objects in the database.