Add setters to all entities and fix incorrect relationship mapping
I have added setter methods to all entities to allow for easy modification of fields.

Entity: I have added setter methods to all entities to allow for easy modification of fields.
Additionally, I have identified and corrected an issue with the relationship mapping between two entities in the code.

Mapping: I have identified and fixed an error in the relationship mapping between two entities.
The addition of setters to all entities and correction of the relationship mapping error will improve the overall functionality and maintainability of the code.