Create SQL backup file, setup React project, and update .gitignore
I have created a backup of the MariaDB data by exporting it to an SQL file and saved it in the sql_backup folder.

Additionally, I have set up the environment for React development and created a project folder named portfolio-react.

To prevent unnecessary files from being tracked in Git, I have updated the .gitignore file to exclude the files and directories that should not be included in the repository. This will ensure that the repository remains clean and organized.

By creating the SQL backup file, I have ensured that the data is saved in case of any unexpected issues with the database. The setup of the React project is a crucial step in the development process, and it will allow us to start building the front-end of the portfolio website.