Add fields to certification model, fix date display bug, and modify design
Add date and description fields to the certification model and modify entity and dto accordingly. Also update the sql_backup.sql file.
Fix a bug where date information was not displaying correctly due to incorrect variable naming.
Modify the design as follows:
a. Change the height of the left panel to 610px.
b. Instead of the content height increasing as more information is added to the right panel, fix the content height and add a scroll bar to the right panel to allow for easy viewing of additional information.
c. When scrolling down in the right panel, make sure the 7 buttons in the left panel follow along.