Upgrade MariaDB version and update portfolio_backup.sql file
While testing the execution method described in the README file, an error occurred due to the MariaDB version being too low to generate the .sql file. To solve this problem, I upgraded the MariaDB version and re-injected the data. As a result, the portfolio_backup.sql file has been updated and the port number in the existing resource.application MariaDB connection settings has been changed from 3307 to 3306. Additionally, due to the change in the project data ID, I modified the React code to display the related project images.